Add files via upload
This projects highlights the usability of python in data cleaning and analysis processes. It also showcases the use of pandas libraries in collecting large amount of data from websites, storing it a structured form, analyzing this data and gaining meaningful insights that can be used in an organization's decision making process.