Add files via upload
This visualization project was created in order to turn complex data into something that’s easier for everyone to understand and work with. 
Visualizations like charts and graphs make it much clearer and easier to understand complex rows of data and speed up decision making process by showing us trends and patterns at a glance.